Combine and simplify AbstractAppender and LeaderAppender and ensure commit futures can be completed during configure/install requests/responses.
